Remove yarn.lock
The yarn.lock file was removed, meaning all the package versions it was locking have been unlocked. This indicates an update to the dependencies, where all packages will be upgraded to the latest version according to the rules in package.json.
Merge pull request #8 from aurelien-oc-mentorship/main

Changing to more consistent tests and integration improvement